Move window analysis to the method
