+++ datafusion/core/src/physical_optimizer/enforce_sorting.rs
@@ -43,9 +43,8 @@ use crate::physical_optimizer::replace_with_order_preserving_variants::{
 };
 use crate::physical_optimizer::sort_pushdown::{pushdown_sorts, SortPushDown};
 use crate::physical_optimizer::utils::{
-    add_sort_above, find_indices, is_coalesce_partitions, is_limit, is_repartition,
-    is_sort, is_sort_preserving_merge, is_sorted, is_union, is_window,
-    merge_and_order_indices, set_difference, unbounded_output, ExecTree,
+    add_sort_above, is_coalesce_partitions, is_limit, is_repartition, is_sort,
+    is_sort_preserving_merge, is_union, is_window, ExecTree,
 };
 use crate::physical_optimizer::PhysicalOptimizerRule;
 use crate::physical_plan::coalesce_partitions::CoalescePartitionsExec;
@@ -56,17 +55,14 @@ use crate::physical_plan::windows::{
 };
 use crate::physical_plan::{with_new_children_if_necessary, Distribution, ExecutionPlan};
 
-use arrow::datatypes::SchemaRef;
 use datafusion_common::tree_node::{Transformed, TreeNode, VisitRecursion};
-use datafusion_common::utils::{get_at_indices, longest_consecutive_prefix};
 use datafusion_common::{plan_err, DataFusionError};
 use datafusion_physical_expr::utils::{
-    convert_to_expr, get_indices_of_matching_exprs, ordering_satisfy,
-    ordering_satisfy_requirement_concrete,
+    ordering_satisfy, ordering_satisfy_requirement_concrete,
 };
-use datafusion_physical_expr::{PhysicalExpr, PhysicalSortExpr, PhysicalSortRequirement};
+use datafusion_physical_expr::{PhysicalSortExpr, PhysicalSortRequirement};
 
-use itertools::{izip, Itertools};
+use itertools::izip;
 
 /// This rule inspects [`SortExec`]'s in the given physical plan and removes the
 /// ones it can prove unnecessary.
@@ -564,106 +560,63 @@ fn analyze_window_sort_removal(
     sort_tree: &mut ExecTree,
     window_exec: &Arc<dyn ExecutionPlan>,
 ) -> Result<Option<PlanWithCorrespondingSort>> {
-    let (window_expr, partition_keys) =
+    let requires_single_partition = matches!(
+        window_exec.required_input_distribution()[sort_tree.idx],
+        Distribution::SinglePartition
+    );
+    let mut window_child =
+        remove_corresponding_sort_from_sub_plan(sort_tree, requires_single_partition)?;
+
+    let (window_expr, new_window) =
         if let Some(exec) = window_exec.as_any().downcast_ref::<BoundedWindowAggExec>() {
-            (exec.window_expr(), &exec.partition_keys)
+            (
+                exec.window_expr(),
+                exec.get_window_for_the_input(&window_child)?,
+            )
         } else if let Some(exec) = window_exec.as_any().downcast_ref::<WindowAggExec>() {
-            (exec.window_expr(), &exec.partition_keys)
+            (
+                exec.window_expr(),
+                exec.get_window_for_the_input(&window_child)?,
+            )
         } else {
             return plan_err!(
                 "Expects to receive either WindowAggExec of BoundedWindowAggExec"
             );
         };
     let partitionby_exprs = window_expr[0].partition_by();
-    let orderby_sort_keys = window_expr[0].order_by();
 
-    // search_flags stores return value of the can_skip_sort.
-    // `None` case represents `SortExec` cannot be removed.
-    // `PartitionSearch` mode stores at which mode executor should work to remove
-    // `SortExec` before it,
-    // `bool` stores whether or not we need to reverse window expressions to remove `SortExec`.
-    let mut search_flags = None;
-    for sort_any in sort_tree.get_leaves() {
-        // Variable `sort_any` will either be a `SortExec` or a
-        // `SortPreservingMergeExec`, and both have a single child.
-        // Therefore, we can use the 0th index without loss of generality.
-        let sort_input = &sort_any.children()[0];
-        let flags = can_skip_sort(partitionby_exprs, orderby_sort_keys, sort_input)?;
-        if flags.is_some() && (search_flags.is_none() || search_flags == flags) {
-            search_flags = flags;
-            continue;
-        }
-        // We can not skip the sort, or window reversal requirements are not
-        // uniform; then sort removal is not possible -- we immediately return.
-        return Ok(None);
-    }
-    let (should_reverse, partition_search_mode) = if let Some(search_flags) = search_flags
-    {
-        search_flags
+    if let Some(new_window) = new_window {
+        Ok(Some(PlanWithCorrespondingSort::new(new_window)))
     } else {
-        // We can not skip the sort return:
-        return Ok(None);
-    };
-    let is_unbounded = unbounded_output(window_exec);
-    if !is_unbounded && partition_search_mode != PartitionSearchMode::Sorted {
-        // Executor has bounded input and `partition_search_mode` is not `PartitionSearchMode::Sorted`
-        // in this case removing the sort is not helpful, return:
-        return Ok(None);
-    };
-
-    let new_window_expr = if should_reverse {
-        window_expr
-            .iter()
-            .map(|e| e.get_reverse_expr())
-            .collect::<Option<Vec<_>>>()
-    } else {
-        Some(window_expr.to_vec())
-    };
-    if let Some(window_expr) = new_window_expr {
-        let requires_single_partition = matches!(
-            window_exec.required_input_distribution()[sort_tree.idx],
-            Distribution::SinglePartition
-        );
-        let mut new_child = remove_corresponding_sort_from_sub_plan(
-            sort_tree,
-            requires_single_partition,
-        )?;
-        let new_schema = new_child.schema();
+        // Exiting input is not sufficient for window
+        let reqs = window_exec
+            .required_input_ordering()
+            .swap_remove(0)
+            .unwrap_or(vec![]);
+        let sort_expr = PhysicalSortRequirement::to_sort_exprs(reqs);
+        // Satisfy the ordering requirement so that window can be run
+        add_sort_above(&mut window_child, sort_expr, None)?;
 
         let uses_bounded_memory = window_expr.iter().all(|e| e.uses_bounded_memory());
-        // If all window expressions can run with bounded memory, choose the
-        // bounded window variant:
-        let new_plan = if uses_bounded_memory {
+        let input_schema = window_child.schema();
+        let new_window = if uses_bounded_memory {
             Arc::new(BoundedWindowAggExec::try_new(
-                window_expr,
-                new_child,
-                new_schema,
-                partition_keys.to_vec(),
-                partition_search_mode,
+                window_expr.to_vec(),
+                window_child,
+                input_schema,
+                partitionby_exprs.to_vec(),
+                PartitionSearchMode::Sorted,
             )?) as _
         } else {
-            if partition_search_mode != PartitionSearchMode::Sorted {
-                // For `WindowAggExec` to work correctly PARTITION BY columns should be sorted.
-                // Hence, if `partition_search_mode` is not `PartitionSearchMode::Sorted` we should convert
-                // input ordering such that it can work with PartitionSearchMode::Sorted (add `SortExec`).
-                // Effectively `WindowAggExec` works only in PartitionSearchMode::Sorted mode.
-                let reqs = window_exec
-                    .required_input_ordering()
-                    .swap_remove(0)
-                    .unwrap_or(vec![]);
-                let sort_expr = PhysicalSortRequirement::to_sort_exprs(reqs);
-                add_sort_above(&mut new_child, sort_expr, None)?;
-            };
             Arc::new(WindowAggExec::try_new(
-                window_expr,
-                new_child,
-                new_schema,
-                partition_keys.to_vec(),
+                window_expr.to_vec(),
+                window_child,
+                input_schema,
+                partitionby_exprs.to_vec(),
             )?) as _
         };
-        return Ok(Some(PlanWithCorrespondingSort::new(new_plan)));
+        Ok(Some(PlanWithCorrespondingSort::new(new_window)))
     }
-    Ok(None)
 }
 
 /// Updates child to remove the unnecessary [`CoalescePartitionsExec`] below it.
@@ -784,152 +737,6 @@ fn get_sort_exprs(
     }
 }
 
-/// Compares physical ordering (output ordering of input executor) with
-/// `partitionby_exprs` and `orderby_keys`
-/// to decide whether existing ordering is sufficient to run current window executor.
-/// A `None` return value indicates that we can not remove the sort in question (input ordering is not
-/// sufficient to run current window executor).
-/// A `Some((bool, PartitionSearchMode))` value indicates window executor can be run with existing input ordering
-/// (Hence we can remove [`SortExec`] before it).
-/// `bool` represents whether we should reverse window executor to remove [`SortExec`] before it.
-/// `PartitionSearchMode` represents, in which mode Window Executor should work with existing ordering.
-fn can_skip_sort(
-    partitionby_exprs: &[Arc<dyn PhysicalExpr>],
-    orderby_keys: &[PhysicalSortExpr],
-    input: &Arc<dyn ExecutionPlan>,
-) -> Result<Option<(bool, PartitionSearchMode)>> {
-    let physical_ordering = if let Some(physical_ordering) = input.output_ordering() {
-        physical_ordering
-    } else {
-        // If there is no physical ordering, there is no way to remove a
-        // sort, so immediately return.
-        return Ok(None);
-    };
-    let orderby_exprs = convert_to_expr(orderby_keys);
-    let physical_ordering_exprs = convert_to_expr(physical_ordering);
-    let equal_properties = || input.equivalence_properties();
-    // indices of the order by expressions among input ordering expressions
-    let ob_indices = get_indices_of_matching_exprs(
-        &orderby_exprs,
-        &physical_ordering_exprs,
-        equal_properties,
-    );
-    if ob_indices.len() != orderby_exprs.len() {
-        // If all order by expressions are not in the input ordering,
-        // there is no way to remove a sort -- immediately return:
-        return Ok(None);
-    }
-    // indices of the partition by expressions among input ordering expressions
-    let pb_indices = get_indices_of_matching_exprs(
-        partitionby_exprs,
-        &physical_ordering_exprs,
-        equal_properties,
-    );
-    let ordered_merged_indices = merge_and_order_indices(&pb_indices, &ob_indices);
-    // Indices of order by columns that doesn't seen in partition by
-    // Equivalently (Order by columns) ∖ (Partition by columns) where `∖` represents set difference.
-    let unique_ob_indices = set_difference(&ob_indices, &pb_indices);
-    if !is_sorted(&unique_ob_indices) {
-        // ORDER BY indices should be ascending ordered
-        return Ok(None);
-    }
-    let first_n = longest_consecutive_prefix(ordered_merged_indices);
-    let furthest_ob_index = *unique_ob_indices.last().unwrap_or(&0);
-    // Cannot skip sort if last order by index is not within consecutive prefix.
-    // For instance, if input is ordered by a,b,c,d
-    // for expression `PARTITION BY a, ORDER BY b, d`, `first_n` would be 2 (meaning a, b defines a prefix for input ordering)
-    // Whereas `furthest_ob_index` would be 3 (column d occurs at the 3rd index of the existing ordering.)
-    // Hence existing ordering is not sufficient to run current Executor.
-    // However, for expression `PARTITION BY a, ORDER BY b, c, d`, `first_n` would be 4 (meaning a, b, c, d defines a prefix for input ordering)
-    // Similarly, `furthest_ob_index` would be 3 (column d occurs at the 3rd index of the existing ordering.)
-    // Hence existing ordering would be sufficient to run current Executor.
-    if first_n <= furthest_ob_index {
-        return Ok(None);
-    }
-    let input_orderby_columns = get_at_indices(physical_ordering, &unique_ob_indices)?;
-    let expected_orderby_columns =
-        get_at_indices(orderby_keys, find_indices(&ob_indices, &unique_ob_indices)?)?;
-    let should_reverse = if let Some(should_reverse) = check_alignments(
-        &input.schema(),
-        &input_orderby_columns,
-        &expected_orderby_columns,
-    )? {
-        should_reverse
-    } else {
-        // If ordering directions are not aligned. We cannot calculate result without changing existing ordering.
-        return Ok(None);
-    };
-
-    let ordered_pb_indices = pb_indices.iter().copied().sorted().collect::<Vec<_>>();
-    // Determine how many elements in the partition by columns defines a consecutive range from zero.
-    let first_n = longest_consecutive_prefix(&ordered_pb_indices);
-    let mode = if first_n == partitionby_exprs.len() {
-        // All of the partition by columns defines a consecutive range from zero.
-        PartitionSearchMode::Sorted
-    } else if first_n > 0 {
-        // All of the partition by columns defines a consecutive range from zero.
-        let ordered_range = &ordered_pb_indices[0..first_n];
-        let input_pb_exprs = get_at_indices(&physical_ordering_exprs, ordered_range)?;
-        let partially_ordered_indices = get_indices_of_matching_exprs(
-            &input_pb_exprs,
-            partitionby_exprs,
-            equal_properties,
-        );
-        PartitionSearchMode::PartiallySorted(partially_ordered_indices)
-    } else {
-        // None of the partition by columns defines a consecutive range from zero.
-        PartitionSearchMode::Linear
-    };
-
-    Ok(Some((should_reverse, mode)))
-}
-
-fn check_alignments(
-    schema: &SchemaRef,
-    physical_ordering: &[PhysicalSortExpr],
-    required: &[PhysicalSortExpr],
-) -> Result<Option<bool>> {
-    let res = izip!(physical_ordering, required)
-        .map(|(lhs, rhs)| check_alignment(schema, lhs, rhs))
-        .collect::<Result<Option<Vec<_>>>>()?;
-    Ok(if let Some(res) = res {
-        if !res.is_empty() {
-            let first = res[0];
-            let all_same = res.into_iter().all(|elem| elem == first);
-            all_same.then_some(first)
-        } else {
-            Some(false)
-        }
-    } else {
-        // Cannot skip some of the requirements in the input.
-        None
-    })
-}
-
-/// Compares `physical_ordering` and `required` ordering, decides whether
-/// alignments match. A `None` return value indicates that current column is
-/// not aligned. A `Some(bool)` value indicates otherwise, and signals whether
-/// we should reverse the window expression in order to avoid sorting.
-fn check_alignment(
-    input_schema: &SchemaRef,
-    physical_ordering: &PhysicalSortExpr,
-    required: &PhysicalSortExpr,
-) -> Result<Option<bool>> {
-    Ok(if required.expr.eq(&physical_ordering.expr) {
-        let physical_opts = physical_ordering.options;
-        let required_opts = required.options;
-        if required.expr.nullable(input_schema)? {
-            let reverse = physical_opts == !required_opts;
-            (reverse || physical_opts == required_opts).then_some(reverse)
-        } else {
-            // If the column is not nullable, NULLS FIRST/LAST is not important.
-            Some(physical_opts.descending != required_opts.descending)
-        }
-    } else {
-        None
-    })
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -941,9 +748,6 @@ mod tests {
         sort_preserving_merge_exec, union_exec,
     };
     use crate::physical_plan::repartition::RepartitionExec;
-    use crate::physical_plan::windows::PartitionSearchMode::{
-        Linear, PartiallySorted, Sorted,
-    };
     use crate::physical_plan::{displayable, Partitioning};
     use crate::prelude::{SessionConfig, SessionContext};
     use crate::test::csv_exec_sorted;
@@ -954,7 +758,6 @@ mod tests {
     use datafusion_expr::JoinType;
     use datafusion_physical_expr::expressions::Column;
     use datafusion_physical_expr::expressions::{col, NotExpr};
-    use datafusion_physical_expr::PhysicalSortExpr;
 
     use crate::physical_optimizer::enforce_distribution::EnforceDistribution;
     use crate::physical_optimizer::utils::get_plan_string;
@@ -984,364 +787,6 @@ mod tests {
         let e = Field::new("e", DataType::Int32, false);
         let schema = Arc::new(Schema::new(vec![a, b, c, d, e]));
         Ok(schema)
-    }
-
-    #[tokio::test]
-    async fn test_is_column_aligned_nullable() -> Result<()> {
-        let schema = create_test_schema()?;
-        let params = vec![
-            ((true, true), (false, false), Some(true)),
-            ((true, true), (false, true), None),
-            ((true, true), (true, false), None),
-            ((true, false), (false, true), Some(true)),
-            ((true, false), (false, false), None),
-            ((true, false), (true, true), None),
-        ];
-        for (
-            (physical_desc, physical_nulls_first),
-            (req_desc, req_nulls_first),
-            expected,
-        ) in params
-        {
-            let physical_ordering = PhysicalSortExpr {
-                expr: col("nullable_col", &schema)?,
-                options: SortOptions {
-                    descending: physical_desc,
-                    nulls_first: physical_nulls_first,
-                },
-            };
-            let required_ordering = PhysicalSortExpr {
-                expr: col("nullable_col", &schema)?,
-                options: SortOptions {
-                    descending: req_desc,
-                    nulls_first: req_nulls_first,
-                },
-            };
-            let res = check_alignment(&schema, &physical_ordering, &required_ordering)?;
-            assert_eq!(res, expected);
-        }
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn test_is_column_aligned_non_nullable() -> Result<()> {
-        let schema = create_test_schema()?;
-
-        let params = vec![
-            ((true, true), (false, false), Some(true)),
-            ((true, true), (false, true), Some(true)),
-            ((true, true), (true, false), Some(false)),
-            ((true, false), (false, true), Some(true)),
-            ((true, false), (false, false), Some(true)),
-            ((true, false), (true, true), Some(false)),
-        ];
-        for (
-            (physical_desc, physical_nulls_first),
-            (req_desc, req_nulls_first),
-            expected,
-        ) in params
-        {
-            let physical_ordering = PhysicalSortExpr {
-                expr: col("non_nullable_col", &schema)?,
-                options: SortOptions {
-                    descending: physical_desc,
-                    nulls_first: physical_nulls_first,
-                },
-            };
-            let required_ordering = PhysicalSortExpr {
-                expr: col("non_nullable_col", &schema)?,
-                options: SortOptions {
-                    descending: req_desc,
-                    nulls_first: req_nulls_first,
-                },
-            };
-            let res = check_alignment(&schema, &physical_ordering, &required_ordering)?;
-            assert_eq!(res, expected);
-        }
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn test_can_skip_ordering_exhaustive() -> Result<()> {
-        let test_schema = create_test_schema3()?;
-        // Columns a,c are nullable whereas b,d are not nullable.
-        // Source is sorted by a ASC NULLS FIRST, b ASC NULLS FIRST, c ASC NULLS FIRST, d ASC NULLS FIRST
-        // Column e is not ordered.
-        let sort_exprs = vec![
-            sort_expr("a", &test_schema),
-            sort_expr("b", &test_schema),
-            sort_expr("c", &test_schema),
-            sort_expr("d", &test_schema),
-        ];
-        let exec_unbounded = csv_exec_sorted(&test_schema, sort_exprs, true);
-
-        // test cases consists of vector of tuples. Where each tuple represents a single test case.
-        // First field in the tuple is Vec<str> where each element in the vector represents PARTITION BY columns
-        // For instance `vec!["a", "b"]` corresponds to PARTITION BY a, b
-        // Second field in the tuple is Vec<str> where each element in the vector represents ORDER BY columns
-        // For instance, vec!["c"], corresponds to ORDER BY c ASC NULLS FIRST, (ordering is default ordering. We do not check
-        // for reversibility in this test).
-        // Third field in the tuple is Option<PartitionSearchMode>, which corresponds to expected algorithm mode.
-        // None represents that existing ordering is not sufficient to run executor with any one of the algorithms
-        // (We need to add SortExec to be able to run it).
-        // Some(PartitionSearchMode) represents, we can run algorithm with existing ordering; and algorithm should work in
-        // PartitionSearchMode.
-        let test_cases = vec![
-            (vec!["a"], vec!["a"], Some(Sorted)),
-            (vec!["a"], vec!["b"], Some(Sorted)),
-            (vec!["a"], vec!["c"], None),
-            (vec!["a"], vec!["a", "b"], Some(Sorted)),
-            (vec!["a"], vec!["b", "c"], Some(Sorted)),
-            (vec!["a"], vec!["a", "c"], None),
-            (vec!["a"], vec!["a", "b", "c"], Some(Sorted)),
-            (vec!["b"], vec!["a"], Some(Linear)),
-            (vec!["b"], vec!["b"], None),
-            (vec!["b"], vec!["c"], None),
-            (vec!["b"], vec!["a", "b"], Some(Linear)),
-            (vec!["b"], vec!["b", "c"], None),
-            (vec!["b"], vec!["a", "c"], Some(Linear)),
-            (vec!["b"], vec!["a", "b", "c"], Some(Linear)),
-            (vec!["c"], vec!["a"], Some(Linear)),
-            (vec!["c"], vec!["b"], None),
-            (vec!["c"], vec!["c"], None),
-            (vec!["c"], vec!["a", "b"], Some(Linear)),
-            (vec!["c"], vec!["b", "c"], None),
-            (vec!["c"], vec!["a", "c"], Some(Linear)),
-            (vec!["c"], vec!["a", "b", "c"], Some(Linear)),
-            (vec!["b", "a"], vec!["a"], Some(Sorted)),
-            (vec!["b", "a"], vec!["b"], Some(Sorted)),
-            (vec!["b", "a"], vec!["c"], Some(Sorted)),
-            (vec!["b", "a"], vec!["a", "b"], Some(Sorted)),
-            (vec!["b", "a"], vec!["b", "c"], Some(Sorted)),
-            (vec!["b", "a"], vec!["a", "c"], Some(Sorted)),
-            (vec!["b", "a"], vec!["a", "b", "c"], Some(Sorted)),
-            (vec!["c", "b"], vec!["a"], Some(Linear)),
-            (vec!["c", "b"], vec!["b"], None),
-            (vec!["c", "b"], vec!["c"], None),
-            (vec!["c", "b"], vec!["a", "b"], Some(Linear)),
-            (vec!["c", "b"], vec!["b", "c"], None),
-            (vec!["c", "b"], vec!["a", "c"], Some(Linear)),
-            (vec!["c", "b"], vec!["a", "b", "c"], Some(Linear)),
-            (vec!["c", "a"], vec!["a"], Some(PartiallySorted(vec![1]))),
-            (vec!["c", "a"], vec!["b"], Some(PartiallySorted(vec![1]))),
-            (vec!["c", "a"], vec!["c"], Some(PartiallySorted(vec![1]))),
-            (
-                vec!["c", "a"],
-                vec!["a", "b"],
-                Some(PartiallySorted(vec![1])),
-            ),
-            (
-                vec!["c", "a"],
-                vec!["b", "c"],
-                Some(PartiallySorted(vec![1])),
-            ),
-            (
-                vec!["c", "a"],
-                vec!["a", "c"],
-                Some(PartiallySorted(vec![1])),
-            ),
-            (
-                vec!["c", "a"],
-                vec!["a", "b", "c"],
-                Some(PartiallySorted(vec![1])),
-            ),
-            (vec!["c", "b", "a"], vec!["a"], Some(Sorted)),
-            (vec!["c", "b", "a"], vec!["b"], Some(Sorted)),
-            (vec!["c", "b", "a"], vec!["c"], Some(Sorted)),
-            (vec!["c", "b", "a"], vec!["a", "b"], Some(Sorted)),
-            (vec!["c", "b", "a"], vec!["b", "c"], Some(Sorted)),
-            (vec!["c", "b", "a"], vec!["a", "c"], Some(Sorted)),
-            (vec!["c", "b", "a"], vec!["a", "b", "c"], Some(Sorted)),
-        ];
-        for (case_idx, test_case) in test_cases.iter().enumerate() {
-            let (partition_by_columns, order_by_params, expected) = &test_case;
-            let mut partition_by_exprs = vec![];
-            for col_name in partition_by_columns {
-                partition_by_exprs.push(col(col_name, &test_schema)?);
-            }
-
-            let mut order_by_exprs = vec![];
-            for col_name in order_by_params {
-                let expr = col(col_name, &test_schema)?;
-                // Give default ordering, this is same with input ordering direction
-                // In this test we do check for reversibility.
-                let options = SortOptions::default();
-                order_by_exprs.push(PhysicalSortExpr { expr, options });
-            }
-            let res =
-                can_skip_sort(&partition_by_exprs, &order_by_exprs, &exec_unbounded)?;
-            // Since reversibility is not important in this test. Convert Option<(bool, PartitionSearchMode)> to Option<PartitionSearchMode>
-            let res = res.map(|(_, mode)| mode);
-            assert_eq!(
-                res, *expected,
-                "Unexpected result for in unbounded test case#: {case_idx:?}, case: {test_case:?}"
-            );
-        }
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn test_can_skip_ordering() -> Result<()> {
-        let test_schema = create_test_schema3()?;
-        // Columns a,c are nullable whereas b,d are not nullable.
-        // Source is sorted by a ASC NULLS FIRST, b ASC NULLS FIRST, c ASC NULLS FIRST, d ASC NULLS FIRST
-        // Column e is not ordered.
-        let sort_exprs = vec![
-            sort_expr("a", &test_schema),
-            sort_expr("b", &test_schema),
-            sort_expr("c", &test_schema),
-            sort_expr("d", &test_schema),
-        ];
-        let exec_unbounded = csv_exec_sorted(&test_schema, sort_exprs, true);
-
-        // test cases consists of vector of tuples. Where each tuple represents a single test case.
-        // First field in the tuple is Vec<str> where each element in the vector represents PARTITION BY columns
-        // For instance `vec!["a", "b"]` corresponds to PARTITION BY a, b
-        // Second field in the tuple is Vec<(str, bool, bool)> where each element in the vector represents ORDER BY columns
-        // For instance, vec![("c", false, false)], corresponds to ORDER BY c ASC NULLS LAST,
-        // similarly, vec![("c", true, true)], corresponds to ORDER BY c DESC NULLS FIRST,
-        // Third field in the tuple is Option<(bool, PartitionSearchMode)>, which corresponds to expected result.
-        // None represents that existing ordering is not sufficient to run executor with any one of the algorithms
-        // (We need to add SortExec to be able to run it).
-        // Some((bool, PartitionSearchMode)) represents, we can run algorithm with existing ordering. Algorithm should work in
-        // PartitionSearchMode, bool field represents whether we should reverse window expressions to run executor with existing ordering.
-        // For instance, `Some((false, PartitionSearchMode::Sorted))`, represents that we shouldn't reverse window expressions. And algorithm
-        // should work in Sorted mode to work with existing ordering.
-        let test_cases = vec![
-            // PARTITION BY a, b ORDER BY c ASC NULLS LAST
-            (vec!["a", "b"], vec![("c", false, false)], None),
-            // ORDER BY c ASC NULLS FIRST
-            (vec![], vec![("c", false, true)], None),
-            // PARTITION BY b, ORDER BY c ASC NULLS FIRST
-            (vec!["b"], vec![("c", false, true)], None),
-            // PARTITION BY a, ORDER BY c ASC NULLS FIRST
-            (vec!["a"], vec![("c", false, true)], None),
-            // PARTITION BY b, ORDER BY c ASC NULLS FIRST
-            (
-                vec!["a", "b"],
-                vec![("c", false, true), ("e", false, true)],
-                None,
-            ),
-            // PARTITION BY a, ORDER BY b ASC NULLS FIRST
-            (vec!["a"], vec![("b", false, true)], Some((false, Sorted))),
-            // PARTITION BY a, ORDER BY a ASC NULLS FIRST
-            (vec!["a"], vec![("a", false, true)], Some((false, Sorted))),
-            // PARTITION BY a, ORDER BY a ASC NULLS LAST
-            (vec!["a"], vec![("a", false, false)], Some((false, Sorted))),
-            // PARTITION BY a, ORDER BY a DESC NULLS FIRST
-            (vec!["a"], vec![("a", true, true)], Some((false, Sorted))),
-            // PARTITION BY a, ORDER BY a DESC NULLS LAST
-            (vec!["a"], vec![("a", true, false)], Some((false, Sorted))),
-            // PARTITION BY a, ORDER BY b ASC NULLS LAST
-            (vec!["a"], vec![("b", false, false)], Some((false, Sorted))),
-            // PARTITION BY a, ORDER BY b DESC NULLS LAST
-            (vec!["a"], vec![("b", true, false)], Some((true, Sorted))),
-            // PARTITION BY a, b ORDER BY c ASC NULLS FIRST
-            (
-                vec!["a", "b"],
-                vec![("c", false, true)],
-                Some((false, Sorted)),
-            ),
-            // PARTITION BY b, a ORDER BY c ASC NULLS FIRST
-            (
-                vec!["b", "a"],
-                vec![("c", false, true)],
-                Some((false, Sorted)),
-            ),
-            // PARTITION BY a, b ORDER BY c DESC NULLS LAST
-            (
-                vec!["a", "b"],
-                vec![("c", true, false)],
-                Some((true, Sorted)),
-            ),
-            // PARTITION BY e ORDER BY a ASC NULLS FIRST
-            (
-                vec!["e"],
-                vec![("a", false, true)],
-                // For unbounded, expects to work in Linear mode. Shouldn't reverse window function.
-                Some((false, Linear)),
-            ),
-            // PARTITION BY b, c ORDER BY a ASC NULLS FIRST, c ASC NULLS FIRST
-            (
-                vec!["b", "c"],
-                vec![("a", false, true), ("c", false, true)],
-                Some((false, Linear)),
-            ),
-            // PARTITION BY b ORDER BY a ASC NULLS FIRST
-            (vec!["b"], vec![("a", false, true)], Some((false, Linear))),
-            // PARTITION BY a, e ORDER BY b ASC NULLS FIRST
-            (
-                vec!["a", "e"],
-                vec![("b", false, true)],
-                Some((false, PartiallySorted(vec![0]))),
-            ),
-            // PARTITION BY a, c ORDER BY b ASC NULLS FIRST
-            (
-                vec!["a", "c"],
-                vec![("b", false, true)],
-                Some((false, PartiallySorted(vec![0]))),
-            ),
-            // PARTITION BY c, a ORDER BY b ASC NULLS FIRST
-            (
-                vec!["c", "a"],
-                vec![("b", false, true)],
-                Some((false, PartiallySorted(vec![1]))),
-            ),
-            // PARTITION BY d, b, a ORDER BY c ASC NULLS FIRST
-            (
-                vec!["d", "b", "a"],
-                vec![("c", false, true)],
-                Some((false, PartiallySorted(vec![2, 1]))),
-            ),
-            // PARTITION BY e, b, a ORDER BY c ASC NULLS FIRST
-            (
-                vec!["e", "b", "a"],
-                vec![("c", false, true)],
-                Some((false, PartiallySorted(vec![2, 1]))),
-            ),
-            // PARTITION BY d, a ORDER BY b ASC NULLS FIRST
-            (
-                vec!["d", "a"],
-                vec![("b", false, true)],
-                Some((false, PartiallySorted(vec![1]))),
-            ),
-            // PARTITION BY b, ORDER BY b, a ASC NULLS FIRST
-            (
-                vec!["a"],
-                vec![("b", false, true), ("a", false, true)],
-                Some((false, Sorted)),
-            ),
-            // ORDER BY b, a ASC NULLS FIRST
-            (vec![], vec![("b", false, true), ("a", false, true)], None),
-        ];
-        for (case_idx, test_case) in test_cases.iter().enumerate() {
-            let (partition_by_columns, order_by_params, expected) = &test_case;
-            let mut partition_by_exprs = vec![];
-            for col_name in partition_by_columns {
-                partition_by_exprs.push(col(col_name, &test_schema)?);
-            }
-
-            let mut order_by_exprs = vec![];
-            for (col_name, descending, nulls_first) in order_by_params {
-                let expr = col(col_name, &test_schema)?;
-                let options = SortOptions {
-                    descending: *descending,
-                    nulls_first: *nulls_first,
-                };
-                order_by_exprs.push(PhysicalSortExpr { expr, options });
-            }
-
-            assert_eq!(
-                can_skip_sort(&partition_by_exprs, &order_by_exprs, &exec_unbounded)?,
-                *expected,
-                "Unexpected result for in unbounded test case#: {case_idx:?}, case: {test_case:?}"
-            );
-        }
-
-        Ok(())
     }
 
     /// Runs the sort enforcement optimizer and asserts the plan
