+++ datafusion/physical-plan/src/windows/bounded_window_agg_exec.rs
@@ -23,7 +23,8 @@
 use crate::expressions::PhysicalSortExpr;
 use crate::metrics::{BaselineMetrics, ExecutionPlanMetricsSet, MetricsSet};
 use crate::windows::{
-    calc_requirements, get_ordered_partition_by_indices, window_ordering_equivalence,
+    calc_requirements, get_ordered_partition_by_indices, get_window_for_the_input,
+    window_ordering_equivalence,
 };
 use crate::{
     ColumnStatistics, DisplayAs, DisplayFormatType, Distribution, ExecutionPlan,
@@ -196,6 +197,19 @@ impl BoundedWindowAggExec {
                 Box::new(LinearSearch::new(ordered_partition_by_indices))
             }
         })
+    }
+
+    /// Constructs either `WindowAggExec` or `BoundedWindowExec` for the given input
+    /// according to specifications of the `window_exprs`.
+    /// `None` represents that with the given input (and its ordering), there is no way to
+    /// construct a window exec. Existing ordering should be changed to be able to run window exec.
+    /// `Some(window exec)` contains the optimal window exec (WindowAggExec or BoundedWindowExec) for the
+    /// given input.
+    pub fn get_window_for_the_input(
+        &self,
+        input: &Arc<dyn ExecutionPlan>,
+    ) -> Result<Option<Arc<dyn ExecutionPlan>>> {
+        get_window_for_the_input(self.window_expr(), input)
     }
 }
